+++ src/domain/comment.md
@@ -1,0 +1,19 @@
+# domain
+
+### 특징
+- 비즈니스 핵심 로직을 포함시킨다.
+- 다른 계층에 의존하지 않는다. (의존성 역전 원칙에 의하여)
+
+### 폴더 구조
+- entities: 모든 도메인의 공통 비즈니스 데이터 모델을 정의할 예정이다. (데이터 구조/타입 정의)
+- special-lecture: 특강 관련 로직 및 규칙 개발
+- user: 사용자 관련 로직 개발
+
+### 각 도메인 로직 내부에 추가될 수 있는 폴더 구조
+- repositories: 조회 및 저장 등 데이터 접근 인터페이스 정의
+- services: 실제 비즈니스 로직 및 규칙 구현 
+- interfaces: 도메인 계약 정의 및 다른 계층과의 통신 규약 정의
+- dto: 요청 DTO 등 실제 계층 간 데이터 전송 객체 정의 및 데이터 검증 로직 정의 (내부 데이터 전송 용도)
+- exceptions: 도메인 별 예외 및 비즈니스 예외 정의
+- events: 도메인 이벤트 핸들러로 실제 로직 완료 이벤트를 정의
+- enums: 도메인 관련 상수 정의
