+++ src/infrastructure/comment.md
@@ -1,0 +1,9 @@
+# infrastructure
+
+### 특징
+- 외부 시스템 연동 로직을 포함시킨다.
+- DB, 외부 API 등 기술적 구현이 여기에 들어간다.
+
+### 폴더 구조
+- database: mysql을 추가할 예정이기 때문에 추가하였다. 데이터베이스 연결과 mysql 구성 파일, 데이터베이스 마이그레이션 등의 로직을 정의할 예정이다. (DB 연결/설정만 담당)
+- repositories: domain에서 정의한 인터페이스를 실제 구현하는 구현체 역할을 하며 실제 데이터베이스 쿼리를 수행할 것이다. (실제 데이터 조작 로직만 담당)
