+++ test/it/example/example.it.spec.ts
@@ -1,4 +1,4 @@
-import { DatabaseModule } from "../../../src/database/database.module";
+import { DatabaseModule } from "../../../src/infrastructure/database/database.module";
 import { Logger } from "@nestjs/common";
 import { Test } from "@nestjs/testing";
